+++ backend/endpoints/tests/test_rom.py
@@ -32,7 +32,7 @@ def test_get_all_roms(access_token, rom, platform):
 
 
 @patch("endpoints.rom.fs_roms_handler.rename_file")
-@patch("endpoints.rom.igdb_handler.get_rom_by_id")
+@patch("endpoints.rom.meta_igdb_handler.get_rom_by_id")
 def test_update_rom(rename_file_mock, get_rom_by_id_mock, access_token, rom):
     response = client.put(
         f"/roms/{rom.id}",
