+++ backend/handler/database/firmware_handler.py
