+++ backend/handler/database/platforms_handler.py
