+++ backend/handler/database/screenshots_handler.py
