+++ backend/handler/filesystem/assets_handler.py
