+++ backend/handler/metadata/__init__.py
@@ -1,5 +1,7 @@
 from .igdb_handler import IGDBBaseHandler
 from .moby_handler import MobyGamesHandler
+from .sgdb_handler import SGDBBaseHandler
 
-igdb_handler = IGDBBaseHandler()
-moby_handler = MobyGamesHandler()
+meta_igdb_handler = IGDBBaseHandler()
+meta_moby_handler = MobyGamesHandler()
+meta_sgdb_handler = SGDBBaseHandler()
